<!--
Hello 👋 Thank you for submitting a pull request.

To help us merge your PR, make sure to follow the instructions below:

- Create or update the tests
- Create or update the documentation at https://github.com/strapi/documentation
- Refer to the issue you are closing in the PR description: Fix #issue
- Specify if the PR is ready to be merged or work in progress (by opening a draft PR)

Please ensure you read the Contributing Guide: https://github.com/strapi/strapi/blob/main/CONTRIBUTING.md
-->

### What does it do?

This is a followup to #14935 to apply the feedback from @madhurisandbhor

It improves performance by using `getByTestId` instead of `getByRole` wherever possible.

|                   | Before | After |
|-------------------|--------|-------|
| index.test.js     | 5.4s  | 5.4s  |
| providers.test.js   | 34.9  | 30.5s |
| plugins.test.js | 66.2s  | 58.1s |
| Total             | 66.8s  | 59.4s |